# docs/openapi.v1.yaml
...
openapi: 3.0.3
info:
  title: MountKey API
  version: 1.0.0
  description: |
    # MountKey Public API
    
    API Platform Data Gunung & Jalur Pendakian Indonesia.
    
    ## Overview
    
    MountKey menyediakan data komprehensif tentang gunung-gunung di Indonesia, 
    termasuk jalur pendakian, checkpoint, dan penilaian risiko cuaca realtime.
    
    ## Authentication
    
    Semua endpoint v1 memerlukan API Key yang valid.
    
    ```
    X-MountKey-API-Key: mk_live_xxxxxxxxxxxxxxxxxxxxxxxx
    ```
    
    Dapatkan API key dengan mendaftar di `/auth/register`.
    
    ## Rate Limiting
    
    Saat ini tidak ada rate limiting. Gunakan API secara bertanggung jawab.
    
    ## Error Handling
    
    Semua error menggunakan format konsisten:
    
    ```json
    {
      "success": false,
      "message": "Error description",
      "error_code": "ERROR_CODE",
      "timestamp": "2026-01-10T06:02:16.000Z"
    }
    ```
    
  contact:
    name: MountKey API Support
    email: api@mountkey.id
    url: https://mountkey.id
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development Server
  - url: https://api.mountkey.id
    description: Production Server (coming soon)
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-MountKey-API-Key
      description: |
        API Key untuk autentikasi.
        Format: `mk_live_xxxxxxxxxxxxxxxxxxxxxxxx`
        
        Dapatkan API key dengan mendaftar di `/auth/register`.

  # ============================================================
  # SCHEMAS
  # ============================================================
  schemas:
...
    Coordinates:
      type: object
      properties:
        latitude:
          type: number
          format: double
          example: -8.1077
        longitude:
          type: number
          format: double
          example: 112.9220
    Pagination:
      type: object
      properties:
        current_page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_items:
          type: integer
          example: 50
        total_pages:
          type: integer
          example: 3
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false
    
    # ----------------------------------------------------------
    # Mountain Schemas
    # ----------------------------------------------------------
    MountainLocation:
      type: object
      properties:
        province:
          type: string
          example: "Jawa Timur"
        regency:
          type: string
          nullable: true
          example: "Lumajang"
        coordinates:
          $ref: '#/components/schemas/Coordinates'
    MountainConservation:
      type: object
      properties:
        area:
          type: string
          nullable: true
          example: "TN Bromo Tengger Semeru"
        type:
          type: string
          enum: [national_park, nature_reserve, wildlife_sanctuary, recreation_park, hunting_park, grand_forest_park, none]
          example: "national_park"
    Mountain:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Gunung Semeru"
        slug:
          type: string
          example: "semeru"
        location:
          $ref: '#/components/schemas/MountainLocation'
        elevation_meters:
          type: integer
          example: 3676
        status:
          type: string
          enum: [active, dormant, extinct]
          example: "active"
        type:
          type: string
          enum: [stratovolcano, shield, caldera, complex, non_volcanic]
          example: "stratovolcano"
        conservation:
          $ref: '#/components/schemas/MountainConservation'
        description:
          type: string
          nullable: true
          example: "Gunung tertinggi di Pulau Jawa"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    MountainDetail:
      allOf:
        - $ref: '#/components/schemas/Mountain'
        - type: object
          properties:
            trail_count:
              type: integer
              example: 2
    MountainSummary:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Gunung Semeru"
        slug:
          type: string
          example: "semeru"
    
    # ----------------------------------------------------------
    # Trail Schemas
    # ----------------------------------------------------------
    TrailBasecamp:
      type: object
      properties:
        name:
          type: string
          example: "Ranu Pane"
        village:
          type: string
          nullable: true
          example: "Ranu Pane"
        district:
          type: string
          nullable: true
          example: "Senduro"
        coordinates:
          $ref: '#/components/schemas/Coordinates'
          nullable: true
    TrailEstimatedHours:
      type: object
      properties:
        up:
          type: number
          format: float
          nullable: true
          example: 8.0
        down:
          type: number
          format: float
          nullable: true
          example: 5.0
        total:
          type: number
          format: float
          nullable: true
          example: 13.0
...
    Trail:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Via Ranu Pane"
        slug:
          type: string
          example: "semeru-via-ranu-pane"
        basecamp:
          $ref: '#/components/schemas/TrailBasecamp'
        distance_km:
          type: number
          format: float
          nullable: true
          example: 18.5
        estimated_hours:
          $ref: '#/components/schemas/TrailEstimatedHours'
        difficulty_level:
          type: integer
          minimum: 1
          maximum: 5
          example: 4
        status:
          type: string
          enum: [open, closed, restricted, maintenance]
          example: "open"
        status_reason:
          type: string
          nullable: true
          example: null
        daily_quota:
          type: integer
          nullable: true
          example: 600
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
        - success
        - message
        - error_code
        - timestamp
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Error description"
        error_code:
          type: string
          example: "ERROR_CODE"
        timestamp:
          type: string
          format: date-time

  # ============================================================
  # RESPONSE SCHEMAS
  # ============================================================
  responses:
    UnauthorizedError:
      description: API Key tidak valid atau tidak ada
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing:
              summary: API Key tidak ada
              value:
                success: false
                message: "API key is required. Provide header: X-MountKey-API-Key"
                error_code: "API_KEY_MISSING"
                timestamp: "2026-01-10T06:02:16.000Z"
            invalid:
              summary: API Key tidak valid
              value:
                success: false
                message: "Invalid API key"
                error_code: "API_KEY_INVALID"
                timestamp: "2026-01-10T06:02:16.000Z"
            expired:
              summary: API Key expired
              value:
                success: false
                message: "API key has expired"
                error_code: "API_KEY_EXPIRED"
                timestamp: "2026-01-10T06:02:16.000Z"
    MountainNotFound:
      description: Gunung tidak ditemukan
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Mountain 'invalid-slug' not found"
            error_code: "MOUNTAIN_NOT_FOUND"
            timestamp: "2026-01-10T06:02:16.000Z"
...
paths:
...
  /v1/mountains:
    get:
      tags:
        - Mountains
      summary: List semua gunung
      description: |
        Mengambil daftar semua gunung dengan pagination dan filtering.
        
        ## Pagination
        
        Gunakan parameter `page` dan `limit` untuk pagination.
        Default: page=1, limit=20, max limit=100.
        
        ## Filtering
        
        - `province`: Filter berdasarkan provinsi
        - `status`: Filter berdasarkan status vulkanik
        
        ## Sorting
        
        - `sort`: Field untuk sorting (name, elevation_meters, province, created_at)
        - `order`: Urutan (asc, desc)
      operationId: listMountains
      parameters:
        - name: page
          in: query
          description: Nomor halaman (default 1)
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Jumlah item per halaman (default 20, max 100)
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: province
          in: query
          description: Filter berdasarkan provinsi
          schema:
            type: string
          example: "Jawa Timur"
        - name: status
          in: query
          description: Filter berdasarkan status vulkanik
          schema:
            type: string
            enum: [active, dormant, extinct]
        - name: sort
          in: query
          description: Field untuk sorting
          schema:
            type: string
            enum: [name, elevation_meters, province, created_at]
            default: name
        - name: order
          in: query
          description: Urutan sorting
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: Daftar gunung berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Mountains retrieved successfully"
                  data:
                    type: object
                    properties:
                      mountains:
                        type: array
                        items:
                          $ref: '#/components/schemas/Mountain'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
                  timestamp:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /v1/mountains/{slug}:
    get:
      tags:
        - Mountains
      summary: Detail gunung
      description: |
        Mengambil detail satu gunung berdasarkan slug.
        
        Response termasuk jumlah trail yang tersedia.
      operationId: getMountainBySlug
      parameters:
        - name: slug
          in: path
          required: true
          description: Slug gunung (URL-friendly identifier)
          schema:
            type: string
          example: semeru
      responses:
        '200':
          description: Detail gunung berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Mountain retrieved successfully"
                  data:
                    type: object
                    properties:
                      mountain:
                        $ref: '#/components/schemas/MountainDetail'
                  timestamp:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/MountainNotFound'
  /v1/mountains/{slug}/trails:
    get:
      tags:
        - Mountains
        - Trails
      summary: Trails untuk gunung
      description: |
        Mengambil semua jalur pendakian untuk gunung tertentu.
      operationId: getMountainTrails
      parameters:
        - name: slug
          in: path
          required: true
          description: Slug gunung
          schema:
            type: string
          example: semeru
      responses:
        '200':
          description: Daftar trails berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Trails retrieved successfully"
                  data:
                    type: object
                    properties:
                      mountain:
                        $ref: '#/components/schemas/MountainSummary'
                      trails:
                        type: array
                        items:
                          $ref: '#/components/schemas/Trail'
                      count:
                        type: integer
                        example: 2
                  timestamp:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/MountainNotFound'
...